# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Lunch API"
  description: "This is a Lunch API"
servers:
  - url: http://0.0.0.0:5000
...
paths:
...
  /api/menus/:
    get:
      tags:
        - "Menus"
      summary: "Get menus information"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenusOut'
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - "Menus"
      summary: "Create new menu"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "1"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: 'Please, specify "restaurant_id","monday", "tuesday", 
                "wednesday", "thursday", "friday", "saturday", "sunday"'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
  /api/menus/{id}:
    get:
      tags:
        - "Menus"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of menu"
          required: true
          schema:
            type: "integer"
      summary: "Get menu by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuOut'
        '404':
          description: "Menu not found"
          content:
            application/json:
              example:
                message: "Menu not found"
    patch:
      security:
        - bearerAuth: [ ]
      tags:
        - "Menus"
      summary: "Update menu's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of menu"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong menu's id"
          content:
            application/json:
              example:
                message: "Menu not found"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - "Menus"
      summary: "Delete menu"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the menu"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong menu's id"
          content:
            application/json:
              example:
                message: "Menu not found"
...
  /api/employees/:
    get:
      tags:
        - "Employees"
      security:
        - bearerAuth: [ ]
      summary: "Get employees information"
      parameters:
        - name: "firstname"
          in: "query"
          description: "Firstname of employee"
          required: false
          schema:
            type: "string"
        - name: "lastname"
          in: "query"
          description: "Lastname of employee"
          required: false
          schema:
            type: "string"
        - name: "email"
          in: "query"
          description: "Email of employee"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - "Employees"
      summary: "Create new employee"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'firstname', 'lastname', 'email', 'password' and 'is_admin'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong input email"
          content:
            application/json:
              example:
                message: "Email {email} already used"
  /api/employees/{id}:
    get:
      tags:
        - "Employees"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          description: "Employee ID"
          required: true
          schema:
            type: "integer"
      summary: "Get employee by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: "Employee not found"
          content:
            application/json:
              example:
                message: "Employee not found"
    patch:
      security:
        - bearerAuth: [ ]
      tags:
        - "Employees"
      summary: "Update user's information"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of employee"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong employee's id"
          content:
            application/json:
              example:
                message: "Employee not found"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - "Employees"
      summary: "Delete employee"
      description: "This can only be done by the logged in user"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the employee"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong employee's id"
          content:
            application/json:
              example:
                message: "Employee not found"
        '405':
          description: "Admin access required"
          content:
            application/json:
              example:
                message: "Not allowed"
  /api/employees/inactive:
    get:
      tags:
        - "Employees"
      security:
        - bearerAuth: [ ]
      summary: "Get inactive employees information"
      description: "This can only be done by the logged in admin"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
  /api/employees/current:
    get:
      tags:
        - "Employees"
      security:
        - bearerAuth: [ ]
      summary: "Get current employee information by email"
      description: "This can only be done by the logged in user"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MenuOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        restaurant_id:
          type: "integer"
          example: 2
        monday:
          type: "string"
          example: "Soup"
        wednesday:
          type: "string"
          example: "Cheese"
        thursday:
          type: "string"
          example: "Salad"
        friday:
          type: "string"
          example: "Tomato"
        saturday:
          type: "string"
          example: "Bread"
        sunday:
          type: "string"
          example: "Cookies"
    MenusOut:
      type: "array"
      items:
        $ref: '#/components/schemas/MenuOut'
    MenuIn:
      type: "object"
      properties:
        restaurant_id:
          type: "integer"
          example: 2
        monday:
          type: "string"
          example: "Soup"
        tuesday:
          type: "string"
          example: "Potato"
        wednesday:
          type: "string"
          example: "Cheese"
        thursday:
          type: "string"
          example: "Salad"
        friday:
          type: "string"
          example: "Tomato"
        saturday:
          type: "string"
          example: "Bread"
        sunday:
          type: "string"
          example: "Cookies"
    UserIn:
      type: "object"
      properties:
        firstname:
          type: "string"
          example: "John"
        lastname:
          type: "string"
          example: "Doe"
        email:
          type: "string"
          example: "test@gmail.com"
        password:
          type: "string"
          example: "password"
        is_admin:
          type: "boolean"
          example: true
...
    UserOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        firstname:
          type: "string"
          example: "John"
        lastname:
          type: "string"
          example: "Doe"
        email:
          type: "string"
          example: "test@gmail.com"
        is_admin:
          type: "boolean"
          example: true
        is_active:
          type: "boolean"
          example: true
    UsersOut:
      type: "array"
      items:
        $ref: '#/components/schemas/UserOut'
...